Remove dtype; cause raising error.

+++ utils/agentreward.py
@@ -167,7 +167,9 @@ class EnvOffline_dist_conf(gym.Env):
         # Gym settings
         self.action_space = spaces.Discrete(self.num_models) 
         # state_dim is 4 , each corresponds to scaled_sc, scaled_thresholds, pred, dist_conf 
-        self.observation_space = spaces.Box(low=0, high=1, shape=(4, ), dtype=np.float32)
+        # remove default dtype
+        # self.observation_space = spaces.Box(low=0, high=1, shape=(4, ), dtype=np.float32)
+        self.observation_space = spaces.Box(low=0, high=1, shape=(4, ))
         self.seed()
         self.reset()
